requirements.txt
langchain-community
python-dotenv
streamlit
langchain-ollama